[IMP]优化core模块
