+++ core/__init__.py
@@ -1,2 +1,2 @@
 # -*- coding: utf-8 -*-
-import core
+import models
