+++ core/__manifest__.py
@@ -2,24 +2,27 @@
 {
     'name': "GOODERP 核心模块",
     'author': "开阖软件",
+    'summary': '隐藏Odoo内置技术复杂性，增加基本权限组',
     'website': "http://www.osbzr.com",
     'category': 'gooderp',
     "description": """
     """,
-    'version': '8.0.0.1',
+    'version': '11.11',
     'depends': ['base',
                 'web_menu_create',
                 'decimal_precision',
-#                'web_readonly_bypass',
-                'web_export_view_good','home_page', 'common_dialog'],
+                'web_export_view_good',
+                'home_page',
+                'common_dialog'],
     'demo': [
-        'core_demo.xml',
+        'data/core_demo.xml',
         ],
     'data': [
         'security/groups.xml',
-        'core_data.xml',
-        'core_view.xml',
         'security/ir.model.access.csv',
+        'data/core_data.xml',
+        'views/core_view.xml',
+        'views/core_templates.xml',
         ],
     'qweb': [
         'static/src/xml/*.xml',
