--- core/core_data.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data noupdate="1">
-        <record id='decimal_quantity' model="decimal.precision">
-            <field name="name">Quantity</field>
-            <field name="digits" eval="1"/>
-        </record>
-        <record id='decimal_amount' model="decimal.precision">
-            <field name="name">Amount</field>
-            <field name="digits" eval="2"/>
-        </record>
-        <record id="warehouse_general" model="warehouse">
-            <field name="name">主仓库</field>
-            <field name="code">General</field>
-            <field name='type'>stock</field>
-        </record>
-        <record id="base.main_company" model="res.company">
-            <field name="name">开阖GoodERP</field>
-            <field name="currency_id" ref="base.CNY" />
-            <field name='cost_method'>fifo</field>
-        </record>
-    </data>
-</openerp>
