+++ core/data/core_demo.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0"?>
 <openerp>
     <data>
+        <!-- 主数据相关的类别，后面会绑定科目 -->
         <record id="customer_category_1" model="core.category">
             <field name="name">一级客户</field>
             <field name="type">customer</field>
@@ -44,7 +45,7 @@
          <record id="credit_card" model="settle.mode">
             <field name="name">信用卡</field>
         </record>
-        
+        <!-- 测试数据用到的客户 -->
         <record id="jd" model="partner">
             <field name="name">京东</field>
             <field name="code">jd</field>
@@ -66,6 +67,7 @@
             <field name="c_category_id" ref="customer_category_vip" />
         </record>
 
+        <!-- 测试数据用到的供应商 -->
         <record id="lenovo" model="partner">
              <field name="name">联想</field>
               <field name="code">lenovo</field>
@@ -80,6 +82,7 @@
             <field name="s_category_id" ref="supplier_category_1"/>
         </record>
 
+        <!-- 付款方式 -->
         <record id="alipay" model="bank.account">
             <field name="name">支付宝(1888888888)</field>
         </record>
@@ -87,10 +90,12 @@
             <field name="name">交通银行(622262 8888888888888)</field>
         </record>
 
+        <!-- 员工 -->
         <record id="lili" model="staff">
         	<field name="name">丽丽</field>
         </record>
 
+        <!-- 服务产品 -->
         <record id="service_1" model="service">
         	<field name="name">咨询服务</field>
         	<field name="get_categ_id" ref="cat_consult"/>
