+++ core/models/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+import clean_data
+import core
+import goods
+import partner
+import pricing
+import res_company
+import res_users
+import warehouse
